Implemented bottom tabbing
