+++ app/(tabs)/_layout.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Tabs } from 'expo-router';
+
+export default function TabLayout() {
+  return (
+    <Tabs>
+        {/* <Tabs.Screen name="Login" /> */}
+    </Tabs>
+  );
+}
