+++ app/inter.js
@@ -1,0 +1,7 @@
+import { Redirect } from "expo-router";
+
+const Inter = () => {
+    return <Redirect href={"/(tabs)/SpotifyRankings"} />;
+};
+
+export default Inter;
